Все файлы были перемещены в папки по их назначению.
