+++ ColorMixerApp/Apps/Info.plist
