+++ ColorMixerApp/Extensions/Float+Extensions.swift
@@ -1,0 +1,14 @@
+//
+//  Float+Extensions.swift
+//  ColorMixerApp
+//
+//  Created by Alexandr Artemov (Mac Mini) on 13.06.2025.
+//
+
+import Foundation
+
+extension Float {
+    func cgFloat() -> CGFloat {
+        CGFloat(self)
+    }
+}
