+++ ColorMixerApp/Storyboards/Base.lproj/LaunchScreen.storyboard
